Don't kill self during shutdown
When Xwayland is running we end up with a client with the same pid
as the compositor in the client list.  We need to avoid killing that
client, as it will interrupt the proper shutdown procedure.

fix T4439

+++ src/bin/e_client.c
@@ -4578,6 +4578,7 @@ e_client_act_kill_begin(E_Client *ec)
    if (!ec->zone) return;
    if (ec->internal) return;
    if (ec->lock_close) return;
+   if (ec->netwm.pid == getpid()) return;
    if ((ec->netwm.pid > 1) && (e_config->kill_process))
      {
         kill(ec->netwm.pid, SIGINT);
